sound: UINT32 instead UIN16 refs #133  svn merge -r 2459:2460 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/branches/yui/support_gimic
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@2574 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ win9x/dialog/d_config.cpp
@@ -165,7 +165,7 @@ void CConfigureDlg::OnOK()
 		update |= SYS_UPDATECFG;
 	}
 
-	UINT16 nSamplingRate;
+	UINT nSamplingRate;
 	if (IsDlgButtonChecked(IDC_RATE11) != BST_UNCHECKED)
 	{
 		nSamplingRate = 11025;
